vibe: Canvas-Firstレイアウトへの移行 (Phase 1)
