+++ index.html
@@ -21,14 +21,14 @@
     <script type="module" src="/src/main.tsx"></script>
     <div id="app-root"></div>
     <div id="main">
-      <div id="header"></div>
-      <div id="canvas-wrapper" style="pointer-events: none">
+      <div id="canvas-wrapper">
         <div id="canvas-overlay"></div>
-        <div id="p5root" style="pointer-events: auto"></div>
+        <div id="p5root"></div>
       </div>
-      <div id="sidebar-right"></div>
-      <div id="footer"></div>
     </div>
+    <div id="toolbar"></div>
+    <div id="poi-panel"></div>
+    <div id="progress-bar"></div>
     <div id="supersampling-overlay"></div>
   </body>
 </html>
