--- src/view/right-sidebar/index.tsx
@@ -1,21 +1,0 @@
-import { useStoreValue } from "@/store/store";
-import { DebugMode } from "./debug-mode/debug-mode";
-import { Operations } from "./operations";
-import { POIHistories } from "./poi-histories";
-
-export const RightSidebar = () => {
-  const isDebugMode = useStoreValue("isDebugMode");
-
-  return (
-    <div className="flex h-full flex-col gap-2">
-      {isDebugMode ? (
-        <DebugMode />
-      ) : (
-        <>
-          <Operations />
-          <POIHistories />
-        </>
-      )}
-    </div>
-  );
-};
